fix legacy issues for cashew/scene and cashew/math

+++ libcashew/include/cashew/graphics/Project.h
@@ -15,31 +15,31 @@ gluLookAt(float eyex, float eyey, float eyez, float centerx,
           float centery, float centerz, float upx, float upy,
           float upz, Matrix4 modelView);
 
-int32
+int32_t
 gluProject(float objx, float objy, float objz,
            const float modelMatrix[16],
            const float projMatrix[16],
-           const int32 viewport[4],
+           const int32_t viewport[4],
            float *winx, float *winy, float *winz);
 
-int32
+int32_t
 gluUnProject(float winx, float winy, float winz,
              const float modelMatrix[16],
              const float projMatrix[16],
-             const int32 viewport[4],
-             float *objx, float *objy, float *objz);
+             const int32_t viewport[4],
+             double *objx, double *objy, double *objz);
 
-int32
+int32_t
 gluUnProject4(float winx, float winy, float winz, float clipw,
               const float modelMatrix[16],
               const float projMatrix[16],
-              const int32 viewport[4],
+              const int32_t viewport[4],
               float nearVal, float farVal,
               float *objx, float *objy, float *objz,
               float *objw);
 
-const int32 GL_TRUE = 1;
-const int32 GL_FALSE = 0;
+const int32_t GL_TRUE = 1;
+const int32_t GL_FALSE = 0;
 };
 };
 #endif
